Fix local nix + coq tests (#359)
`tox -e coqXX-py36` is supposed to use Nix to run the Coq tests locally with
version `XX`, but at some point, this was broken by changing how Coqtail
locates Coq executables. The CI didn't catch it because it uses `tox -e
coq-py36`, which just uses whatever version is currently installed.

The solution is to have the test harness pass the `COQBIN` environment variable
to `Coqtop.start`.